Update Clang build to use libc++abi and Ninja.
This patch does two things:

(A) It switches the Clang build over to using Ninja instead of Make.
Ninja is faster, and more importantly, automatically determines the
number of jobs best for the machine. This allows the Clang build to
go faster.

(B) Use libc++abi instead of libstdc++ when building libc++. I'm
not exactly sure why libstdc++ was being used initially, but I suspect
it was due to old bugs in libc++abi's demangler. Assuming that is the
case, it should be safe to switch over now.